being anal

+++ src/lint/ast.c
@@ -17,7 +17,7 @@ ast_node_t* ast_get_child(ast_node_t* node, uint32_t i) {
     return child_node;
 }
 
-size_t ast_number_of_children(ast_node_t* node) {
+uint16_t ast_number_of_children(ast_node_t* node) {
     if (node->children == NULL) {
 	return 0;
     }
