+++ src/lint/ast.h
@@ -180,7 +180,7 @@ typedef struct ast_node {
 
 const char* ast_node_name_to_string(node_name_t name);
 ast_node_t* ast_get_child(ast_node_t* node, uint32_t i);
-size_t ast_number_of_children(ast_node_t* node);
+uint16_t ast_number_of_children(ast_node_t* node);
 ast_node_t* ast_last_child(ast_node_t* node);
 void ast_print(ast_node_t* node, uint16_t level);
