+++ src/lint/equations.h
@@ -8,7 +8,7 @@ typedef dynarray_t equations_t;
 
 void equations_init(equations_t* equations);
 void equations_append(equations_t* equations, equation_t* equation);
-equation_t* equations_get(equations_t* equations, size_t i);
+equation_t* equations_get(equations_t* equations, uint16_t i);
 void equations_print(equations_t* equations);
 
 #endif
