+++ src/lint/types.h
@@ -8,6 +8,6 @@ typedef dynarray_t types_t;
 
 types_t* types_new(void);
 void types_add(types_t* types, type_t* type);
-type_t* types_get(types_t* types, size_t i);
+type_t* types_get(types_t* types, uint16_t i);
 
 #endif
